remove to agent example folder
